Reuse InsetLocator to make twinned axes follow their parents.
From a positioning PoV, a twinned axes is just like an inset axes whose
position exactly matches the parent's position.  Doing so removes the
need for the heuristic in `Figure.subplots_adjust`/`GridSpec.update`
where a non-gridspec-managed Axes would track the position of any Axes
with which it shared either xaxis or yaxis, which was a proxy for
twinning (per 721b949).  This would cause incorrect behavior in rare
cases such as
```python
from pylab import *
ax = subplot(221)
axes([.6, .6, .3, .3], sharex=ax)
subplots_adjust(left=0)
```
where the `subplots_adjust` call would make the second axes go on top of
the first.
