+++ lib/matplotlib/figure.py
@@ -2424,15 +2424,7 @@ default: 'top'
                                  "constrained_layout==False. ")
         self.subplotpars.update(left, bottom, right, top, wspace, hspace)
         for ax in self.axes:
-            if not isinstance(ax, SubplotBase):
-                # Check if sharing a subplots axis
-                if isinstance(ax._sharex, SubplotBase):
-                    ax._sharex.update_params()
-                    ax.set_position(ax._sharex.figbox)
-                elif isinstance(ax._sharey, SubplotBase):
-                    ax._sharey.update_params()
-                    ax.set_position(ax._sharey.figbox)
-            else:
+            if isinstance(ax, SubplotBase):
                 ax.update_params()
                 ax.set_position(ax.figbox)
         self.stale = True
